# packages/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: CyberPulse AVA API - Enterprise Cybersecurity Companion Platform
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /devices:
    get:
      operationId: getDevices
      tags: [devices]
      summary: List all managed devices with compliance status
      parameters:
        - name: complianceState
          in: query
          schema:
            type: string
            enum: [compliant, noncompliant, unknown, all]
            default: all
        - name: platform
          in: query
          schema:
            type: string
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
      responses:
        "200":
          description: List of managed devices
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Device"
  /devices/{id}:
    get:
      operationId: getDevice
      tags: [devices]
      summary: Get device details
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Device detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Device"
        "404":
          description: Device not found
...
  /alerts:
    get:
      operationId: getAlerts
      tags: [alerts]
      summary: Get security alert feed
      parameters:
        - name: severity
          in: query
          schema:
            type: string
            enum: [critical, high, medium, low, informational, all]
            default: all
        - name: status
          in: query
          schema:
            type: string
            enum: [active, resolved, dismissed, all]
            default: active
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
      responses:
        "200":
          description: Security alert feed
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/SecurityAlert"
  /alerts/{id}:
    get:
      operationId: getAlert
      tags: [alerts]
      summary: Get alert details
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Alert detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SecurityAlert"
        "404":
          description: Alert not found
  /alerts/{id}/status:
    patch:
      operationId: updateAlertStatus
      tags: [alerts]
      summary: Update alert status (resolve or dismiss)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AlertStatusUpdate"
      responses:
        "200":
          description: Updated alert
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SecurityAlert"
        "404":
          description: Alert not found
...
  /assistant/chat:
    post:
      operationId: sendAssistantMessage
      tags: [assistant]
      summary: Send a message to the AI Security Assistant
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AssistantMessageInput"
      responses:
        "200":
          description: Assistant response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AssistantResponse"
  /assistant/history:
    get:
      operationId: getAssistantHistory
      tags: [assistant]
      summary: Get chat history with the AI Security Assistant
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
      responses:
        "200":
          description: Chat history
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AssistantMessage"

  # ─── Teams Notifications ─────────────────────────────────────────────────────
  /notifications/preferences:
    get:
      operationId: getNotificationPreferences
      tags: [notifications]
      summary: Get Teams notification preferences
      responses:
        "200":
          description: Notification preferences
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NotificationPreferences"
    put:
      operationId: updateNotificationPreferences
      tags: [notifications]
      summary: Update Teams notification preferences
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/NotificationPreferences"
      responses:
        "200":
          description: Updated notification preferences
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NotificationPreferences"
components:
  schemas:
...
    Device:
      type: object
      required: [id, deviceName, platform, osVersion, complianceState, lastSyncDateTime, userId, userDisplayName]
      properties:
        id:
          type: string
        deviceName:
          type: string
        platform:
          type: string
        osVersion:
          type: string
        complianceState:
          type: string
          enum: [compliant, noncompliant, unknown, inGracePeriod]
        lastSyncDateTime:
          type: string
        userId:
          type: string
        userDisplayName:
          type: string
        userEmail:
          type: string
        managedBy:
          type: string
        encryptionEnabled:
          type: boolean
        firewallEnabled:
          type: boolean
        antivirusEnabled:
          type: boolean
        osUpdateStatus:
          type: string
          enum: [upToDate, updateAvailable, updateRequired, unknown]
        nonComplianceReasons:
          type: array
          items:
            type: string
...
    SecurityAlert:
      type: object
      required: [id, title, severity, status, source, category, createdAt]
      properties:
        id:
          type: string
        title:
          type: string
        description:
          type: ["string", "null"]
        severity:
          type: string
          enum: [critical, high, medium, low, informational]
        status:
          type: string
          enum: [active, resolved, dismissed]
        source:
          type: string
          enum: [defender, entra, intune, darktrace, pentera, manual]
        category:
          type: string
        affectedEntity:
          type: ["string", "null"]
        remediationSteps:
          type: ["string", "null"]
        createdAt:
          type: string
        resolvedAt:
          type: ["string", "null"]
    AlertStatusUpdate:
      type: object
      required: [status]
      properties:
        status:
          type: string
          enum: [resolved, dismissed]
        resolutionNote:
          type: string

    # ─── Darktrace ─────────────────────────────────────────────────────────────
...
    AssistantMessageInput:
      type: object
      required: [message]
      properties:
        message:
          type: string
        context:
          type: string
          enum: [general, alerts, devices, training, score]
    AssistantMessage:
      type: object
      required: [id, role, content, createdAt]
      properties:
        id:
          type: string
        role:
          type: string
          enum: [user, assistant]
        content:
          type: string
        createdAt:
          type: string
    AssistantResponse:
      type: object
      required: [message, suggestedActions]
      properties:
        message:
          $ref: "#/components/schemas/AssistantMessage"
        suggestedActions:
          type: array
          items:
            type: string

    # ─── Notifications ────────────────────────────────────────────────────────
    NotificationPreferences:
      type: object
      required: [teamsEnabled, alertsEnabled, scoreChangesEnabled, weeklyDigestEnabled]
      properties:
        teamsEnabled:
          type: boolean
        teamsWebhookUrl:
          type: ["string", "null"]
        alertsEnabled:
          type: boolean
        alertMinSeverity:
          type: string
          enum: [critical, high, medium, low]
        scoreChangesEnabled:
          type: boolean
        weeklyDigestEnabled:
          type: boolean
        weeklyDigestDay:
          type: string
          enum: [monday, tuesday, wednesday, thursday, friday]