# packages/api-zod/src/generated/types/notificationPreferences.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * CyberPulse AVA API - Enterprise Cybersecurity Companion Platform
 * OpenAPI spec version: 0.1.0
 */
import type { NotificationPreferencesAlertMinSeverity } from './notificationPreferencesAlertMinSeverity';
import type { NotificationPreferencesWeeklyDigestDay } from './notificationPreferencesWeeklyDigestDay';

export interface NotificationPreferences {
  teamsEnabled: boolean;
  /** @nullable */
  teamsWebhookUrl?: string | null;
  alertsEnabled: boolean;
  alertMinSeverity?: NotificationPreferencesAlertMinSeverity;
  scoreChangesEnabled: boolean;
  weeklyDigestEnabled: boolean;
  weeklyDigestDay?: NotificationPreferencesWeeklyDigestDay;
}